split geo-api
