+++ project/geo_api/views/__init__.py
@@ -1,0 +1,3 @@
+from .feature_viewset import FeatureViewSet
+
+__all__ = ["FeatureViewSet"]
